requirements.txt
codetiming==1.4.0
matplotlib==3.10.0
numpy==2.2.3
RPi.GPIO==0.7.1
smbus2==0.5.0
